Fix approveclaim email

+++ api/src/controllers/pitch.controller.ts
@@ -9,7 +9,7 @@ import {
   sendDeclinedPitchMail,
 } from '../mail/sender';
 import { populatePitch } from '../populators';
-import { PitchService, UserService } from '../services';
+import { PitchService, TeamService, UserService } from '../services';
 import { isWriterOrEditor } from '../services/pitch.service';
 import { sendFail, sendNotFound, sendSuccess } from '../utils/helpers';
 import { extractOptions, extractPopulateQuery } from './utils';
@@ -291,6 +291,12 @@ export const approveClaimRequest = async (
   const { userId, teamId } = req.body;
   const { writer, editor } = req.query;
 
+  const team = await TeamService.getOne(teamId);
+  if (!team) {
+    sendNotFound(res, `Team with id ${teamId} not found`);
+    return;
+  }
+
   let pitch = await PitchService.removeTeamFromPendingContributors(
     req.params.id,
     userId,
@@ -327,7 +333,7 @@ export const approveClaimRequest = async (
     'default',
   )) as BasePopulatedPitch;
 
-  sendClaimRequestApprovedMail(user, defaultPopulatedPitch, req.user);
+  sendClaimRequestApprovedMail(user, defaultPopulatedPitch, req.user, team);
 
   sendSuccess(res, 'Claim approved successfully', defaultPopulatedPitch);
 };
